Add TIME_MULTIPLIER
Current numbers in scenarios and everywhere were assumed
to be in seconds, except latencies. This made
the simulations not correct as latencies usually
are expected to be wihin a second range and not
in a multiples of seconds.

However simulating milliseconds is wasteful as it's x1000
multiplier. Thus TIME_MULTIPLIER was introduced with
the initial value of 10 which allows to leave numbers
as they are, in seconds.

This is not a proper solution and just merely a quick solution
for the problem to better model the situation, as previously
latency wasn't modelled properly.

Signed-off-by: Lukas Pukenis <lukas.pukenis@gmail.com>