Time is not a global anymore
The reason why the charts didn't make sense
was because singular instance of time was used
which inherently is exactly what we want to achieve.

So now in a hacky way time was placed on each device
and it's starting value simulated randomly.

This now shows charts more like it where local batching
aligns only local events and when enabling
incoming batching, all events seem to align better.

Receive and local batchings are implemented by using threshold
this is constantly halved. This eliminates signal tending towards the divider.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>