New batch strategy
Batching has been reworked and now works as such:
- local batching batches locally on deadline-threshold
however it only does so at signal deadline. This simplifies
things a lot as the previous method was to check if the
current time falls into the allowed range, effectively
always shortening the period towards period-threshold.

An alternative could be to find the latest point in time
for emission which current methods effectively does.

- remote batching triggers the check as well, very simply.

TLDR: On every signal's deadline we try to gather as much
events as possible and we also trigger this check on receiving data.